chore(apprise): moved all Apprise stuffs into apprise_plugin
