+++ changedetectionio/apprise_plugin/assets.py
